Add workaround for xml keys appearing differently in xmlops vs qb
In the CompanyRet in particular some boolean keys are defined by
the xmlops as "isX" while QB actually sends "IsX". This is a pretty
specific workaround because it's unknown if the variant used by
the xmlops appears in the wild, if it's just incorrect, or if this
should actually be case insensitive.